# docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cars Management API",
    "description": "## Introduction\n\nThis project is a implementation of a Car Management API in [Node.js](https://nodejs.org/en/). It uses [Express](https://expressjs.com/) as web framework and [PostgreSQL](https://www.postgresql.org/) as database.\n\nThis application allows new users to register and authenticate using the JSON web token. Registered member can view the list of available cars . Admin has additional privileges such as adding new cars to databases or other actions to manage cars & users.",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/users/api/v1": {
      "get": {
        "tags": ["Users"],
        "summary": "Get Users List",
        "description": "This endpoint is useful for displaying a list of available users, resource can only be accessed by admin & super admin",
        "operationId": "GetUsersList",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2003"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully got users list",
                        "data": [
                          {
                            "id": "06811c73-77c1-45e5-9f52-0172838551da",
                            "name": "member",
                            "email": "member5@gmail.com",
                            "role": "member"
                          },
                          {
                            "id": "bb9d4447-5bb2-4738-99c9-57d5bbd11627",
                            "name": "member",
                            "email": "member1@gmail.com",
                            "role": "member"
                          },
                          {
                            "id": "f3e8837c-3e05-4764-b402-132dce787545",
                            "name": "member",
                            "email": "member2@gmail.com",
                            "role": "member"
                          },
                          {
                            "id": "54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                            "name": "member",
                            "email": "member@gmail.com",
                            "role": "member"
                          },
                          {
                            "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
                            "name": "supermin",
                            "email": "supermin@gmail.com",
                            "role": "super admin"
                          }
                        ]
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully got users list",
                  "data": [
                    {
                      "id": "06811c73-77c1-45e5-9f52-0172838551da",
                      "name": "member",
                      "email": "member5@gmail.com",
                      "role": "member"
                    },
                    {
                      "id": "bb9d4447-5bb2-4738-99c9-57d5bbd11627",
                      "name": "member",
                      "email": "member1@gmail.com",
                      "role": "member"
                    },
                    {
                      "id": "f3e8837c-3e05-4764-b402-132dce787545",
                      "name": "member",
                      "email": "member2@gmail.com",
                      "role": "member"
                    },
                    {
                      "id": "54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                      "name": "member",
                      "email": "member@gmail.com",
                      "role": "member"
                    },
                    {
                      "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
                      "name": "supermin",
                      "email": "supermin@gmail.com",
                      "role": "super admin"
                    }
                  ]
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create User",
        "description": "This endpoint is use to create a user, resource can only be accessed by super admin.",
        "operationId": "CreateUser",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "name",
                  "email",
                  "password",
                  "confirmPassword",
                  "role"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "admin tamvan"
                  },
                  "email": {
                    "type": "string",
                    "example": "admin2@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "pass123"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "pass123"
                  },
                  "role": {
                    "type": "string",
                    "example": "admin"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2001"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully created user",
                        "data": {
                          "id": "6748c8f1-cf70-468c-aa27-216902add8e5",
                          "name": "admin tamvan",
                          "email": "admin@gmail.com",
                          "role": "admin"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully created user",
                  "data": {
                    "id": "6748c8f1-cf70-468c-aa27-216902add8e5",
                    "name": "admin tamvan",
                    "email": "admin@gmail.com",
                    "role": "admin"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/api/v1/getCurrentUser": {
      "get": {
        "tags": ["Users"],
        "summary": "Get Current User",
        "description": "This endpoint is use for displaying recent users",
        "operationId": "GetCurrentUser",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2001"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully got current user with auth cridential : eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNhYjVlNDBiLTc5OTktNGUyYS04OTE2LWE5NTJmMjkwMGJlZiIsIm5hbWUiOiJzdXBlcm1pbiIsImVtYWlsIjoic3VwZXJtaW5AZ21haWwuY29tIiwicm9sZSI6InN1cGVyIGFkbWluIiwiaWF0IjoxNjY2MzMyNzEwLCJleHAiOjE2NjY0MTkxMTB9.DN-9BdG7OeVnx-8iEuf8_Yd8bLZhRYT3Af6mhwdZhm4",
                        "data": {
                          "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
                          "name": "supermin",
                          "email": "supermin@gmail.com",
                          "role": "super admin"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully got current user with auth cridential : eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjNhYjVlNDBiLTc5OTktNGUyYS04OTE2LWE5NTJmMjkwMGJlZiIsIm5hbWUiOiJzdXBlcm1pbiIsImVtYWlsIjoic3VwZXJtaW5AZ21haWwuY29tIiwicm9sZSI6InN1cGVyIGFkbWluIiwiaWF0IjoxNjY2MzMyNzEwLCJleHAiOjE2NjY0MTkxMTB9.DN-9BdG7OeVnx-8iEuf8_Yd8bLZhRYT3Af6mhwdZhm4",
                  "data": {
                    "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
                    "name": "supermin",
                    "email": "supermin@gmail.com",
                    "role": "super admin"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/api/v1/f3e8837c-3e05-4764-b402-132dce78754": {
      "get": {
        "tags": ["Users"],
        "summary": "Get User By ID",
        "description": "This endpoint is use for display a user by id, resource can only be accessed by admin & super admin.",
        "operationId": "GetUserByID",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2001"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully got user with id : f3e8837c-3e05-4764-b402-132dce787545",
                        "data": {
                          "id": "f3e8837c-3e05-4764-b402-132dce787545",
                          "name": "member",
                          "email": "member2@gmail.com",
                          "role": "member"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully got user with id : f3e8837c-3e05-4764-b402-132dce787545",
                  "data": {
                    "id": "f3e8837c-3e05-4764-b402-132dce787545",
                    "name": "member",
                    "email": "member2@gmail.com",
                    "role": "member"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/api/v1/54dbda24-1fe9-4b5e-a133-80e7f06cb7c1": {
      "put": {
        "tags": ["Users"],
        "summary": "Update User",
        "description": "This endpoint is use for update a user by id, resource can only be accessed by super admin.\\\\",
        "operationId": "UpdateUser",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "encoding": {},
              "schema": {
                "required": [
                  "name",
                  "email",
                  "password",
                  "confirmPassword",
                  "role"
                ],
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "updated Member"
                  },
                  "email": {
                    "type": "string",
                    "example": "memberupdater@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "pass123"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "pass123"
                  },
                  "role": {
                    "type": "string",
                    "example": "member"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m2001"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully updated user with id : 54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                        "data": {
                          "id": "54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                          "name": "updated Member",
                          "email": "memberupdater@gmail.com",
                          "role": "member"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully updated user with id : 54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                  "data": {
                    "id": "54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
                    "name": "updated Member",
                    "email": "memberupdater@gmail.com",
                    "role": "member"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/api/v1/06811c73-77c1-45e5-9f52-0172838551daza": {
      "delete": {
        "tags": ["Users"],
        "summary": "Remove User",
        "description": "This endpoint is use for remove user by id, resource can only be accessed by super admin.",
        "operationId": "RemoveUser",
        "parameters": [],
        "responses": {
          "default": {
            "description": "",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/m4041"
                    },
                    {
                      "example": {
                        "status": "success",
                        "message": "Successfully removed user with id : ccbb575a-450c-499c-ad1f-3e374c9ab826"
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Successfully removed user with id : ccbb575a-450c-499c-ad1f-3e374c9ab826"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Data": {
        "title": "Data",
        "required": ["id", "name", "email", "role"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "example": {
          "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
          "name": "supermin",
          "email": "supermin@gmail.com",
          "role": "super admin"
        }
      },
      "m4041": {
        "title": "m4041",
        "required": ["status", "message"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "status": "error",
          "message": "Username or password is incorrect"
        }
      },
      "m2001": {
        "title": "m2001",
        "required": ["status", "message", "data"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Data"
          }
        },
        "example": {
          "status": "success",
          "message": "Successfully sign up !",
          "data": {
            "id": "06811c73-77c1-45e5-9f52-0172838551da",
            "name": "member",
            "email": "member@gmail.com",
            "role": "member"
          }
        }
      },
...
      "m2003": {
        "title": "m2003",
        "required": ["status", "message", "data"],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Data"
            },
            "description": ""
          }
        },
        "example": {
          "status": "success",
          "message": "Successfully got users list",
          "data": [
            {
              "id": "06811c73-77c1-45e5-9f52-0172838551da",
              "name": "member",
              "email": "member5@gmail.com",
              "role": "member"
            },
            {
              "id": "bb9d4447-5bb2-4738-99c9-57d5bbd11627",
              "name": "member",
              "email": "member1@gmail.com",
              "role": "member"
            },
            {
              "id": "f3e8837c-3e05-4764-b402-132dce787545",
              "name": "member",
              "email": "member2@gmail.com",
              "role": "member"
            },
            {
              "id": "54dbda24-1fe9-4b5e-a133-80e7f06cb7c1",
              "name": "member",
              "email": "member@gmail.com",
              "role": "member"
            },
            {
              "id": "3ab5e40b-7999-4e2a-8916-a952f2900bef",
              "name": "supermin",
              "email": "supermin@gmail.com",
              "role": "super admin"
            }
          ]
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}